+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Compound date intervals are joined more naturally, such as "3 hours, 5 minutes and 6 seconds" instead of "3 hours 5 minutes 6 seconds".
 - Rename `Change` rule to `ChangeInterval`.
 - Rename `Date` constraint to `DateInterval`.
+- The dates of former passwords must be immutable.
 
 ### Fixed
 - Nothing